requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.49
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
textblob==0.17.1
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
mysql-connector-python==9.0.0
# Rasa is large; ensure compatible version is installed in your environment
rasa==3.6.18
# Voice processing dependencies
speechrecognition==3.10.0
pydub==0.25.1
pyttsx3==2.90
pyaudio==0.2.11
